// frontend/src/openapi/compassClient/models/SystemStatusDto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * OpenAPI definition
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface SystemStatusDto
 */
export interface SystemStatusDto {
    /**
     * 
     * @type {string}
     * @memberof SystemStatusDto
     */
    commitId?: string;
    /**
     * 
     * @type {boolean}
     * @memberof SystemStatusDto
     */
    backendIsReachable?: boolean;
    /**
     * 
     * @type {boolean}
     * @memberof SystemStatusDto
     */
    databaseIsReachable?: boolean;
    /**
     * 
     * @type {boolean}
     * @memberof SystemStatusDto
     */
    auth0IsReachable?: boolean;
}

/**
 * Check if a given object implements the SystemStatusDto interface.
 */
export function instanceOfSystemStatusDto(value: object): boolean {
    return true;
}

export function SystemStatusDtoFromJSON(json: any): SystemStatusDto {
    return SystemStatusDtoFromJSONTyped(json, false);
}

export function SystemStatusDtoFromJSONTyped(json: any, ignoreDiscriminator: boolean): SystemStatusDto {
    if (json == null) {
        return json;
    }
    return {
        
        'commitId': json['commitId'] == null ? undefined : json['commitId'],
        'backendIsReachable': json['backendIsReachable'] == null ? undefined : json['backendIsReachable'],
        'databaseIsReachable': json['databaseIsReachable'] == null ? undefined : json['databaseIsReachable'],
        'auth0IsReachable': json['auth0IsReachable'] == null ? undefined : json['auth0IsReachable'],
    };
}

export function SystemStatusDtoToJSON(value?: SystemStatusDto | null): any {
    if (value == null) {
        return value;
    }
    return {
        
        'commitId': value['commitId'],
        'backendIsReachable': value['backendIsReachable'],
        'databaseIsReachable': value['databaseIsReachable'],
        'auth0IsReachable': value['auth0IsReachable'],
    };
}